Move HttpError, RequestParser into files
