+++ src/actions/start/httperror.ts
@@ -1,0 +1,5 @@
+export class HttpError extends Error {
+  constructor(public readonly status: number, text: string, public readonly expose: boolean = true) {
+    super(text);
+  }
+}
