+++ tslint.json
@@ -13,6 +13,7 @@
     "no-class": false,
     "no-console": false,
     "no-delete": true,
+    "no-empty": false,
     "no-empty-interfaces": false,
     "no-expression-statement": false,
     "no-if-statement": false,
